requirements/dev.txt
-r base.txt
django-debug-toolbar==4.2.0

# For Django Extensions ex: run_server_plus to run on https locally
django-extensions==3.2.3
Werkzeug==3.0.1
pyOpenSSL==23.3.0

pytest==7.4.4
pytest-django==4.7.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
pytest-cov==4.1.0

# For user switching
django-hijack==3.4.5

# Linting libraries
flake8==7.0.0
black==24.2.0
pylint-django==2.5.5
